fixed solar_input

+++ solar_input.py
@@ -51,11 +51,11 @@ def parse_star_parameters(line, star):
     line = line.split()
     star.R = int(line[1])
     star.color = line[2]
-    star.m = int(line[3])
-    star.x = int(line[4])
-    star.y = int(line[5])
-    star.Vx = int(line[6])
-    star.Vy = int(line[7])
+    star.m = float(line[3])
+    star.x = float(line[4])
+    star.y = float(line[5])
+    star.Vx = float(line[6])
+    star.Vy = float(line[7])
 
 def parse_planet_parameters(line, planet):
     """Считывает данные о планете из строки.
@@ -75,11 +75,11 @@ def parse_planet_parameters(line, planet):
     line = line.split()
     planet.R = int(line[1])
     planet.color = line[2]
-    planet.m = int(line[3])
-    planet.x = int(line[4])
-    planet.y = int(line[5])
-    planet.Vx = int(line[6])
-    planet.Vy = int(line[7])
+    planet.m = float(line[3])
+    planet.x = float(line[4])
+    planet.y = float(line[5])
+    planet.Vx = float(line[6])
+    planet.Vy = float(line[7])
 
 
 def write_space_objects_data_to_file(output_filename, space_objects):
